requirements.txt
einops==0.8.0 
matplotlib==3.7.0 
numpy==1.24.4 
timm==1.0.11 
plotly 
tensorboard 
hydra-core 
ipykernel 
rich 
pytest 
scikit-learn 
torchmetrics==1.6.2 
transformers
wandb
open_clip_torch
sentencepiece